Move logger stuff out of Config
Also added util/fs.hpp, which defines directoryExists and fileExists
