+++ include/3dstris/config.hpp
@@ -55,10 +55,7 @@ struct Config {
 	const FS_Path configPath =
 		fsMakePath(PATH_ASCII, "/3ds/3dstris/config.json");
 
-    static constexpr auto LOG_PATH = "sdmc:/3ds/3dstris/log.log";
-    const FS_Path logPath = fsMakePath(PATH_ASCII, "/3ds/3dstris/log.log");
-
-    Games games;
+	Games games;
 	L10n l10n;
 	Keybinds keybinds;
