+++ src/config/keybinds.cpp
@@ -3,17 +3,9 @@
 #include <rapidjson/filewritestream.h>
 #include <rapidjson/writer.h>
 #include <sds.h>
-
 #include <3dstris/config/keybinds.hpp>
+#include <3dstris/util/fs.hpp>
 #include <3dstris/util/log.hpp>
-
-static bool fileExists(const FS_Archive archive, const FS_Path& path) {
-	Handle handle;
-
-	return R_SUCCEEDED(
-			   FSUSER_OpenFile(&handle, archive, path, FS_OPEN_READ, 0)) &&
-		   R_SUCCEEDED(FSFILE_Close(handle));
-}
 
 static bool validateJson(const rapidjson::Document& doc) {
 	return !doc.HasParseError() && doc.IsObject();
